Feat: La pizarra ya carga

+++ app/src/main/java/com/example/gestionpisoscompartidos/model/dtos/PointDeltaDTO.kt
@@ -1,8 +1,8 @@
 package com.example.gestionpisoscompartidos.model.dtos
 
 data class PointDeltaDTO(
-    val x: Short,
-    val y: Short,
+    val x: Float,
+    val y: Float,
     val size: Float,
     var color: Byte,
 )
